Change temperature to decimal

+++ ancillary/database/create_tables.sql
@@ -5,8 +5,12 @@ use iaq;
 create table cores (id INT NOT NULL AUTO_INCREMENT, core_name VARCHAR(24), name VARCHAR(20), tz VARCHAR(40), 
 UNIQUE(core_name), UNIQUE(name), PRIMARY KEY (id));
 
-create table readings (id INT NOT NULL AUTO_INCREMENT, temperature INT, humidity INT, 
-dust INT, sewer INT, hcho INT, 
+create table readings (id INT NOT NULL AUTO_INCREMENT, 
+temperature DECIMAL(4,2), 
+humidity INT, 
+dust INT, 
+sewer INT, 
+hcho INT, 
 core_id INT, ts INT UNSIGNED, PRIMARY KEY (id), UNIQUE(core_id, ts));
 
 create table locations (id INT NOT NULL AUTO_INCREMENT,
